zephyr: Refactor display user_data handling
We need to associate more state with the display to implement
efficient handling of EPD displays. Create a lvgl_disp_data struct and
store a pointer to that struct in disp_drv->user_data. This struct
currently only stores a pointer to the display device and a local copy
of its capabilities.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>