zephyr: allow placing LVGL rendering buffers into custom section
Some applications may want to locate LVGL rendering buffers in a custom
memory location, such as tightly coupled or external memory. To
facilitate this, add a Kconfig CONFIG_LV_Z_VBD_CUSTOM_SECTION, which
when set will locate the LVGL rendering buffers in a section labelled as
".lvgl_buf". This section can then be located in a custom memory region
by the application.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>