zephyr: Add a mapping of device to lv_indev_t
Add a general mechanism for creating a mapping of the zephyr `device`
struct to lvgls `indev` type. All information for the creation
of the `indev` is stored in an iterable struct section. Generally one
entry in said section is created per mappable `device`. Currently the
exception to that rule is the touch input device, for which only one
instance is created according to the chosen `zephyr,keyboard-scan` or
`zephyr,touch-input` node.

Signed-off-by: Fabian Blatz <fabianblatz@gmail.com>