fix: ifdef function to avoid unused function build warning
This function is only used if LV_DRAW_SW_SUPPORT_RGB565_SWAPPED is
set, let's ifdef it to avoid a build warning in Zephyr.

Note this commit is in a branch that forks from the upstream master
branch. It is a provisional patch to fix Zephyr's CI.
This branch should be discarded once the real fix is merged in
lvgl's master branch.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>